requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# REQUIREMENTS FROM MEAN SHIFTED AD USING CL
faiss-gpu==1.7.1

joblib==1.0.1
numpy==1.21
Pillow==9.0.0
pkg-resources==0.0.0
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.12.0
tqdm==4.61.0
typing-extensions==3.10.0.0



# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger

diagvibsix
setuptools==60.9.3

imageio>=2.8.0
